# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: SafeSchool safeguarding and incident reporting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools/{schoolId}/pupils:
    get:
      operationId: listPupilsBySchool
      tags: [users]
      summary: List pupils for a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: string
        - name: className
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Pupil list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PupilSummary"
  /schools/{schoolId}/staff:
    get:
      operationId: listStaffBySchool
      tags: [users]
      summary: List staff for a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}/avatar:
    post:
      operationId: updateAvatar
      tags: [users]
      summary: Update user avatar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvatarBody"
      responses:
        "200":
          description: Avatar updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarResponse"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        role:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        yearGroup:
          type: ["string", "null"]
        className:
          type: ["string", "null"]
        avatarType:
          type: ["string", "null"]
        avatarValue:
          type: ["string", "null"]
        avatarImageUrl:
          type: ["string", "null"]
        parentOf:
          type: array
          items:
            type: string
        active:
          type: boolean
        lastLogin:
          type: ["string", "null"]
      required:
        - id
        - schoolId
        - role
        - firstName
        - lastName
        - active
...
    PupilSummary:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        yearGroup:
          type: ["string", "null"]
        className:
          type: ["string", "null"]
        avatarType:
          type: ["string", "null"]
        avatarValue:
          type: ["string", "null"]
      required:
        - id
        - firstName
        - lastName
    UpdateAvatarBody:
      type: object
      properties:
        avatarType:
          type: string
        avatarValue:
          type: string
      required:
        - avatarType
        - avatarValue
    AvatarResponse:
      type: object
      properties:
        avatarType:
          type: string
        avatarValue:
          type: string
        avatarImageUrl:
          type: ["string", "null"]
      required:
        - avatarType
        - avatarValue
...